implemented pagination to doctors repository

+++ src/domain/doctors/repository/doctors_repository_trait.rs
@@ -8,8 +8,8 @@ pub trait DoctorsRepositoryTrait {
     async fn create_doctor(&self, doctor: NewDoctor) -> anyhow::Result<()>;
     async fn get_doctors(
         &self,
-        page: Option<i16>,
-        page_size: Option<i16>,
+        page: Option<i64>,
+        page_size: Option<i64>,
     ) -> anyhow::Result<Vec<Doctor>>;
     async fn get_doctor_by_id(&self, doctor_id: Uuid) -> anyhow::Result<Doctor>;
 }
